enterprise/lifecycle: allow multiple rules to apply to a single object (and thus, multiple concurrent reviews)
